Add more custom attributes and docs
Add more custom attributes for all float vector variants.

Move `EffectProperties` into the `properties` module.

Make the `properties` and `attributes` modules public, and add some
module level documentation for both features.

Fixes #262
